chatbot_tutorial/requirements.txt
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.5.0
python-dotenv>=1.0.0
streamlit>=1.30.0
gradio>=4.0.0
flask>=3.0.0
requests>=2.31.0
tiktoken>=0.5.0
